Moving file system utils to Wasp.Cli.FileSystem
